upgrade module
