+++ migrate.py
@@ -1,5 +1,4 @@
 import os, sys
-from openerp import tools
 import glob
 import importlib
 
@@ -10,18 +9,9 @@ import config
 # TODO: rename to something else that is not migrate since that sounds like migration to v8
 # TODO: tests
 # TODO: Readme , example config file
-
-
-def install_module(cr, pool, module_name):
-    module_pool = pool.get('ir.module.module')
-    module_pool.update_list(cr, 1)
-    module = module_pool.search(cr, 1, [('name', '=', module_name)])
-    module = module[0]
-    module = module_pool.browse(cr, 1, module)
-    if module.state != "installed":
-        module_pool.button_immediate_install(cr, 1, [module.id])
-        tools.config['db_name']
-        pool.restart()
+# TODO: what happens for conflicting migrations
+# TODO: integration test
+# TODO: update module
 
 
 def get_steps():
